Move validate_self_client implementation into ValidationContext
