+++ crates/ibc/src/mock/mod.rs
@@ -11,8 +11,6 @@ pub mod header;
 #[cfg(any(test, feature = "mocks"))]
 pub mod host;
 #[cfg(any(test, feature = "mocks"))]
-pub mod host_helper;
-#[cfg(any(test, feature = "mocks"))]
 pub mod ics18_relayer;
 #[cfg(any(test, feature = "mocks", feature = "mocks-no-std"))]
 pub mod misbehaviour;
